pyproject.toml
[project]
name = "fastapi-worker-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "loguru>=0.7.3",
    "notebook>=7.5.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.10",
    "ty>=0.0.9",
    "uvicorn[standard]>=0.40.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
